split apu channels into separate files
makes it much more clean and easier to focus on one channel, rather than scrolling through a bunch of audio channel code that you dont need
